# api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LanguageLearn API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/study/session": {
      "post": {
        "tags": [
          "study"
        ],
        "summary": "Create Study Session",
        "operationId": "create_study_session_v1_study_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/study/review": {
      "post": {
        "tags": [
          "study"
        ],
        "summary": "Submit Review",
        "operationId": "submit_review_v1_study_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/progress/{course_id}": {
      "get": {
        "tags": [
          "study"
        ],
        "summary": "Get Course Progress",
        "operationId": "get_course_progress_v1_progress__course_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/progress": {
      "get": {
        "tags": [
          "study"
        ],
        "summary": "Get All Progress",
        "description": "Return progress for every course from precalculated cache.",
        "operationId": "get_all_progress_v1_progress_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllProgressResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/review-schedule/{course_id}": {
      "get": {
        "tags": [
          "study"
        ],
        "summary": "Get Review Schedule",
        "description": "Return full SRS detail for all concepts the user has started in a course.",
        "operationId": "get_review_schedule_v1_review_schedule__course_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllProgressResponse": {
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseProgressResponse"
            },
            "type": "array",
            "title": "Courses"
          }
        },
        "type": "object",
        "required": [
          "courses"
        ],
        "title": "AllProgressResponse"
      },
...
      "CourseProgressResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "levels": {
            "items": {
              "$ref": "#/components/schemas/CefrProgressItem"
            },
            "type": "array",
            "title": "Levels"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "levels"
        ],
        "title": "CourseProgressResponse"
      },
...
      "ExerciseType": {
        "type": "string",
        "enum": [
          "forward_mc",
          "reverse_mc",
          "cloze",
          "reverse_cloze",
          "forward_typing",
          "reverse_typing"
        ],
        "title": "ExerciseType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewRequest": {
        "properties": {
          "concept_id": {
            "type": "string",
            "format": "uuid",
            "title": "Concept Id"
          },
          "rating": {
            "type": "string",
            "pattern": "^(again|hard|good|easy)$",
            "title": "Rating"
          },
          "exercise_type": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "review_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "rating",
          "exercise_type",
          "correct"
        ],
        "title": "ReviewRequest"
      },
      "ReviewResponse": {
        "properties": {
          "concept_id": {
            "type": "string",
            "format": "uuid",
            "title": "Concept Id"
          },
          "new_forward_difficulty": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "forward_consecutive_correct": {
            "type": "integer",
            "title": "Forward Consecutive Correct"
          },
          "new_reverse_difficulty": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "reverse_consecutive_correct": {
            "type": "integer",
            "title": "Reverse Consecutive Correct"
          },
          "is_mastered": {
            "type": "boolean",
            "title": "Is Mastered"
          },
          "fsrs_due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fsrs Due"
          },
          "difficulty_advanced": {
            "type": "boolean",
            "title": "Difficulty Advanced"
          },
          "mastery_changed": {
            "type": "boolean",
            "title": "Mastery Changed"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "new_forward_difficulty",
          "forward_consecutive_correct",
          "new_reverse_difficulty",
          "reverse_consecutive_correct",
          "is_mastered",
          "fsrs_due",
          "difficulty_advanced",
          "mastery_changed"
        ],
        "title": "ReviewResponse"
      },
      "ReviewScheduleResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConceptProgressDetail"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "items"
        ],
        "title": "ReviewScheduleResponse"
      },
      "StudySessionItem": {
        "properties": {
          "concept_id": {
            "type": "string",
            "format": "uuid",
            "title": "Concept Id"
          },
          "exercise_type": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "is_review": {
            "type": "boolean",
            "title": "Is Review"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "target_text": {
            "type": "string",
            "title": "Target Text"
          },
          "concept_type": {
            "$ref": "#/components/schemas/ConceptType"
          },
          "cefr_level": {
            "$ref": "#/components/schemas/CefrLevel"
          },
          "exercise_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercise Id"
          },
          "exercise_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercise Data"
          },
          "correct_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Answer"
          },
          "distractors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distractors"
          },
          "sentence_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentence Template"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "exercise_type",
          "is_review",
          "source_text",
          "target_text",
          "concept_type",
          "cefr_level"
        ],
        "title": "StudySessionItem"
      },
      "StudySessionRequest": {
        "properties": {
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "session_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Session Size",
            "default": 20
          },
          "concept_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Ids"
          }
        },
        "type": "object",
        "required": [
          "course_id"
        ],
        "title": "StudySessionRequest"
      },
      "StudySessionResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudySessionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_due_reviews": {
            "type": "integer",
            "title": "Total Due Reviews"
          },
          "new_concepts_added": {
            "type": "integer",
            "title": "New Concepts Added"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_due_reviews",
          "new_concepts_added"
        ],
        "title": "StudySessionResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}